% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{analyticalThompsonSampling}
\alias{analyticalThompsonSampling}
\title{Analytical Thompson shares}
\usage{
analyticalThompsonSampling(dist_cache)
}
\arguments{
\item{dist_cache}{List containing the characteristics of the distributions of treatments.}
}
\value{
List that contains (i) the analytical Thompson shares and (ii) the absolute error associated to the numerical integration procedure.
TODO: be clear about the distribution. Only implemented for bernoulli.
}
\description{
Get the analytical Thompson shares based on the characteristics of the distributions of treatments.
}
