% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{bayesianUpdateBinomial}
\alias{bayesianUpdateBinomial}
\title{Bayesian update for a Binomial outcome}
\usage{
bayesianUpdateBinomial(treatment, outcome, alpha0 = 1, beta0 = 1)
}
\arguments{
\item{treatment}{A vector.}

\item{outcome}{A vector.}

\item{alpha0}{A vector.}

\item{beta0}{A vector.}
}
\value{
A list with the distribution and parameters of the posterior distribution of theta.
}
\description{
Obtain the posterior of the parameter theta for an outcome distributed as a Binomial with prior Beta.
}
