% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{bayesianUpdatePoisson}
\alias{bayesianUpdatePoisson}
\title{Bayesian update for a Poisson outcome}
\usage{
bayesianUpdatePoisson(treatment, outcome, alpha0 = 1, beta0 = 0)
}
\arguments{
\item{treatment}{A vector.}

\item{outcome}{A vector.}

\item{alpha0}{A vector.}

\item{beta0}{A vector.}
}
\value{
A list with the distribution and parameters of the posterior distribution of theta.
}
\description{
Obtain the posterior of the parameter theta for an outcome distributed as poisson with prior Gamma.
}
