% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{BinomialExplorationSampling}
\alias{BinomialExplorationSampling}
\title{Full exploration sampling method}
\usage{
BinomialExplorationSampling(
  treatment,
  outcome,
  alpha0 = 1,
  beta0 = 1,
  sample_resolution = 1e+05,
  treatment_cost = NA
)
}
\arguments{
\item{treatment}{A vector.}

\item{outcome}{A vector.}

\item{alpha0}{A vector.}

\item{beta0}{A vector.}

\item{sample_resolution}{A number.}

\item{treatment_cost}{A vector.}
}
\value{
A vector with the proportion shares, obtained through exploration sampling, for each treatment.
}
\description{
Get exploration sampling shares from a sample of outcomes and treatments.
}
