% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{cdfProductBeta}
\alias{cdfProductBeta}
\title{Product of 1 Beta pdf and k-1 Beta cdfs}
\usage{
cdfProductBeta(x, dist_cache, index)
}
\arguments{
\item{x}{Value at which the pdfs and cdfs are evaluated.}

\item{dist_cache}{List containing the characteristics of the distributions of treatments.}

\item{index}{Scalar denoting the treatment associated to the beta pdf.}
}
\value{
product of the beta pdf and k-1 beta cdfs.
}
\description{
Calculate the product of 1 Beta pdf and k-1 Beta cdfs, where k is the number of treatments. The selected treatment that characterizes the
Beta pdf is declared with the argument index. The k-1 beta cdfs are calculated using the distributions of the remaining treatments.
}
