% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consilium.R
\name{checkTimeWithinWindow}
\alias{checkTimeWithinWindow}
\title{Check message answers within a window}
\usage{
checkTimeWithinWindow(agg_twilio_data, twilio_data, HOUR_THRESHOLD)
}
\arguments{
\item{agg_twilio_data}{A data.frame with the aggregated dataset of twilio logs at id level.}

\item{twilio_data}{A data.frame with logs from the twilio application.}

\item{HOUR_THRESHOLD}{A scalar. Number of tolerance hours.}
}
\value{
A data.frame with the final outcome. (=1 if student answered within HOUR_THRESHOLD hours)
}
\description{
Generate a data.frame with the time difference between sending a message and receiving an answer from the student. Those
that take more than HOUR_THRESHOLD hours answering o never answered have an outcome = 0.
}
