% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{encodeLabels}
\alias{encodeLabels}
\title{Encode treatment labels}
\usage{
encodeLabels(treatment, treatment_encoding)
}
\arguments{
\item{treatment}{A vector.}

\item{treatment_encoding}{A dataframe.}
}
\value{
A vector with the treatment transformed based on values of the column 'new' in the
treatment_encoding dataframe.
}
\description{
Transform treatment vector values based on the treatment_encoding dataframe from 'original' to 'new'.
}
