% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{evaluateDiffThompsonShares}
\alias{evaluateDiffThompsonShares}
\title{Difference between Thompson shares and analytical Thompson shares}
\usage{
evaluateDiffThompsonShares(thompson_shares, analytical_thompson_shares)
}
\arguments{
\item{thompson_shares}{Vector containing Thompson shares.}

\item{analytical_thompson_shares}{List containing the analytical Thompson shares and their associated absolute error.}
}
\value{
Minimum difference between Thomson shares and analytical Thompson shares.
}
\description{
Get the difference between Thompson shares and analytical Thompson shares.
}
