% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{expectedRegret}
\alias{expectedRegret}
\title{Expected regret}
\usage{
expectedRegret(shares, true_treatment_effects)
}
\arguments{
\item{shares}{A vector.}

\item{true_treatment_effects}{A vector.}
}
\value{
A number equal to the expected policy regret.
}
\description{
Generate the expected policy regret of the sample allocation based on p-shares and true treatment effects.
}
