% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{explorationSampling}
\alias{explorationSampling}
\title{Exploration sampling}
\usage{
explorationSampling(prop_shares)
}
\arguments{
\item{prop_shares}{A vector.}
}
\value{
A vector with the proportion shares, obtained through exploration sampling, for each treatment.
}
\description{
Get exploration sampling shares by modifying the proportion shares obtained through Thompson sampling.
}
