% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{generateBinomialOutcome}
\alias{generateBinomialOutcome}
\title{Simulate a binomial outcome}
\usage{
generateBinomialOutcome(single_treatment, true_theta)
}
\arguments{
\item{single_treatment}{A number between 1 and the number of treatments.}

\item{true_theta}{An ordered vector with the probabilities of success of each treatment. Its length must be equal to the number of treatments.}
}
\value{
A simulated binomial outcome.
}
\description{
Simulate a single binomial outcome based on a vector of true treatment effects.
}
