% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{generateBinomialOutcomeTable}
\alias{generateBinomialOutcomeTable}
\title{Outcome table for a Binomial outcome variable}
\usage{
generateBinomialOutcomeTable(
  history_df,
  dist_cache,
  thompson_pshares,
  treatment_encoding,
  wave_index
)
}
\arguments{
\item{history_df}{a dataframe.}

\item{dist_cache}{A list.}

\item{thompson_pshares}{A vector.}

\item{treatment_encoding}{A dataframe.}

\item{wave_index}{A scalar.}
}
\value{
A dataframe showing main outcomes for a Binomial outcome variable
}
\description{
Generate the outcome table for a Binomial outcome table. Columns are:
\itemize{
\item Treatment
\item Treatment label
\item Success
\item Attempts
\item Average outcome
\item Mean of the posterior distribution
\item Standard deviation of the posterior distribution
\item Thompson shares
}
}
