% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{generateEncoding}
\alias{generateEncoding}
\title{Treatment encoding}
\usage{
generateEncoding(treatment, sort_treat = T)
}
\arguments{
\item{treatment}{A vector.}

\item{sort_treat}{A logical.}
}
\value{
A dataframe with the original treatment encoding on column 'original' and the new treatment encoding on
column 'new'.
}
\description{
Generate a treatment encoding based on the values on the vector treatment. This is usually the first step in most
the procedures on this package.
}
