% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consilium.R
\name{generateExSampProps}
\alias{generateExSampProps}
\title{Update bayesian posterior and generate exploration sampling shares}
\usage{
generateExSampProps(
  treat_data,
  twilio_data,
  WAVE_INDEX,
  HOUR_THRESHOLD,
  treatment_encoding
)
}
\arguments{
\item{treat_data}{A data.frame containing (i) ids, (ii) treatment, (iii) wave and (iv) phone_number.}

\item{twilio_data}{A data.frame with logs from the twilio application.}

\item{WAVE_INDEX}{a scalar. Index of the current wave.}

\item{HOUR_THRESHOLD}{A scalar. Number of tolerance hours.}
}
\value{
A list with (i) a vector containing the Exploration Sampling shares, (ii) a list with the characteristics of the
updated posterior, (iii) a data.frame with the current (and previous) wave information.
}
\description{
This function (i) updates the Bayesian posterior based on the results from the wave and (ii) generates Exploration Sampling
}
