% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consilium.R
\name{getNotTreatedObs}
\alias{getNotTreatedObs}
\title{Identify not treated observations}
\usage{
getNotTreatedObs(treat_ids, twilio_ids)
}
\arguments{
\item{treat_ids}{A vector with ids from treat_data dataframe.}

\item{twilio_ids}{A vector with ids from the Twilio dataframe.}
}
\value{
A vector with ids that were not treated.
}
\description{
Identify not treated observations in the Twilio dataset. These observations appear on the treat_data dataframe,
but they are not in the Twilio dataframe. This suggests that observations were not treated even though there
was a treatment allocated to them.
}
