% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{inSampleRegret}
\alias{inSampleRegret}
\title{In-sample regret}
\usage{
inSampleRegret(treatment, true_treatment_effects, sample_size)
}
\arguments{
\item{treatment}{A vector.}

\item{true_treatment_effects}{A vector.}

\item{sample_size}{A scalar}
}
\value{
A number equal to the expected policy regret.
}
\description{
Generate the in-sample regret of the sample allocation based on p-shares and true treatment effects.
}
