% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{plotAverageOutcome}
\alias{plotAverageOutcome}
\title{Plot average outcomes}
\usage{
plotAverageOutcome(
  history_df,
  treatment_encoding,
  n_waves,
  y_limits = c(0, 1),
  cumulative = T,
  font_size = 20
)
}
\arguments{
\item{history_df}{A dataframe with the complete sample at observation level.}

\item{treatment_encoding}{A dataframe with the treatment encoding.}

\item{n_waves}{A scalar for the number of waves.}

\item{y_limits}{A vector with two values for the limits of y in the plot.}

\item{cumulative}{A logical. True if the quantities are cumulative.}

\item{font_size}{A scalar for the font size.}
}
\value{
A ggplot of the evolution of average outcomes across waves.
}
\description{
Plot evolution of average outcomes (no bayesian estimation) across waves.
}
