% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{plotAverageOutcomePerWave}
\alias{plotAverageOutcomePerWave}
\title{Plot average outcome (bar plot)}
\usage{
plotAverageOutcomePerWave(hisotry_df, treatment_encoding, font_size = 20)
}
\arguments{
\item{treatment_encoding}{A dataframe with the treatment encoding.}

\item{font_size}{A scalar for the font size.}

\item{history_df}{A dataframe with the complete sample at observation level.}
}
\value{
A ggplot (bar plot) of the average outcomes at each wave (not cumulative).
}
\description{
Generate a bar plot of the average outcome of each treatment at each wave (not cumulative).
}
