% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{plotPosteriorDistributions}
\alias{plotPosteriorDistributions}
\title{Plot posterior distributions}
\usage{
plotPosteriorDistributions(theta_matrix, treatment_encoding, font_size = 20)
}
\arguments{
\item{theta_matrix}{A matrix from dist_cache with the theta drawns from the posterioir distribution.}

\item{treatment_encoding}{A dataframe with the treatment encoding.}

\item{font_size}{A scalar for the font size.}
}
\value{
A ggplot of posterior density distributions for a single wave.
}
\description{
Plot posterior density distributions for a single wave.
}
