% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{plotPosteriorMean}
\alias{plotPosteriorMean}
\title{Plot posterior mean and HPI}
\usage{
plotPosteriorMean(
  summary_posterior,
  n_waves,
  y_limits = c(0, 1),
  font_size = 20
)
}
\arguments{
\item{summary_posterior}{A dataframe with information of the posterior distributions across waves.}

\item{n_waves}{A scalar for the number of waves.}

\item{y_limits}{A vector with two values for the limits of y in the plot.}

\item{font_size}{A scalar for the font size.}
}
\value{
A ggplot of the evolution of the posterior means and HPIs across waves.
}
\description{
Plot the evolution of the posterior mean and the 95\% Highest Density Interval (HPI) across waves.
}
