% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{plotProportionalShares}
\alias{plotProportionalShares}
\title{Plot Exploration Sampling shares}
\usage{
plotProportionalShares(exploration_props_df, font_size = 20)
}
\arguments{
\item{exploration_props_df}{A dataframe with information about the exploration sampling across waves.}

\item{font_size}{A scalar for the font size.}
}
\value{
A ggplot of the evolution of Exploration Sampling shares.
}
\description{
Plot evolution of Exploration Sampling shares across waves.
}
