% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{proportionalAssignment}
\alias{proportionalAssignment}
\title{Sample based on proportions}
\usage{
proportionalAssignment(shares, sample_size)
}
\arguments{
\item{shares}{A vector.}

\item{sample_size}{A number.}
}
\value{
A vector with the new sample based on the proportions in the vector shares.
}
\description{
Generate a sample of size sample_size based on the proportions in the vector shares.
}
