% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consilium.R
\name{removeExtraObs}
\alias{removeExtraObs}
\title{Remove irrelevant observations from Twilio dataset}
\usage{
removeExtraObs(twilio_data_raw, original_ids)
}
\arguments{
\item{twilio_data_raw}{A dataframe with the raw Twilio data.}

\item{original_ids}{A vector with ids from the treat_data dataframe.}
}
\value{
A dataframe with the twilio_data_raw information with irrelevant observations removed.
}
\description{
Remove observations from the Twilio dataset whose id is not in original_ids. The vector original_ids
is often obtained from the treat_data dataset, which contains the treatment allocation of the interventions.
Observations removed did not have a treatment allocated to them, but appear in the Twilio dataset.
}
