% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampleDistribution}
\alias{sampleDistribution}
\title{Sample from a posterior distribution.}
\usage{
sampleDistribution(dist_cache, sample_resolution = 50000)
}
\arguments{
\item{dist_cache}{A list with the structure of a bayesianUpdateBinomial output.}

\item{sample_resolution}{A number.}
}
\value{
A matrix with the samples drawn from the posterior distribution.
}
\description{
Sample from a posterior distribution.
}
