% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{thompsonSampling}
\alias{thompsonSampling}
\title{Thompson sampling}
\usage{
thompsonSampling(theta_matrix, treatment_cost = NA)
}
\arguments{
\item{theta_matrix}{A matrix.}

\item{treatment_cost}{A vector.}
}
\value{
A vector with the proportion shares, obtained through Thompson sampling, for each treatment.
}
\description{
Use Thompson sampling to find the proportion shares of each treatment.
}
